Moved line drawing code out of time-series-graph so that it can be used by both the hydrograph-svg and brush-svg
