+++ assets/src/styles/components/hydrograph/_app.scss
@@ -172,8 +172,6 @@
 }
 
 .slider-wrapper {
-  height: 0;
-
   #cursor-slider {
     max-width: none;
     position: relative;
